Reset texture pack

+++ src/minecraft/org/getspout/spout/packet/PacketTexturePack.java
@@ -4,6 +4,11 @@ import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.File;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.src.RenderEngine;
+
+import org.getspout.spout.client.SpoutClient;
 import org.getspout.spout.io.FileUtil;
 import org.getspout.spout.texture.TexturePackAction;
 import org.getspout.spout.io.Download;
@@ -36,17 +41,26 @@ public class PacketTexturePack implements SpoutPacket{
 
 	@Override
 	public void run(int PlayerId) {
-		String fileName = FileUtil.getFileName(url);
-		if (!FileUtil.isZippedFile(fileName)) {
-			System.out.println("Rejecting Invalid Texture Pack: " + fileName);
-			return;
+		if (url.equals("[none]")) {
+			if (SpoutClient.getHandle().renderEngine.oldPack != null) {
+				RenderEngine rengine = SpoutClient.getHandle().renderEngine;
+				rengine.texturePack.setTexturePack(rengine.oldPack);
+				rengine.oldPack = null;
+				rengine.refreshTextures();
+			}
+		} else {
+			String fileName = FileUtil.getFileName(url);
+			if (!FileUtil.isZippedFile(fileName)) {
+				System.out.println("Rejecting Invalid Texture Pack: " + fileName);
+				return;
+			}
+			File texturePack = new File(FileUtil.getTexturePackDirectory(), fileName);
+			if (texturePack.exists()) {
+				texturePack.delete();
+			}
+			Download download = new Download(fileName, FileUtil.getTexturePackDirectory(), url, new TexturePackAction(fileName, FileUtil.getTexturePackDirectory()));
+			FileDownloadThread.getInstance().addToDownloadQueue(download);
 		}
-		File texturePack = new File(FileUtil.getTexturePackDirectory(), fileName);
-		if (texturePack.exists()) {
-			texturePack.delete();
-		}
-		Download download = new Download(fileName, FileUtil.getTexturePackDirectory(), url, new TexturePackAction(fileName, FileUtil.getTexturePackDirectory()));
-		FileDownloadThread.getInstance().addToDownloadQueue(download);
 	}
 
 	@Override
